+++ src/pantr/bspline/_bspline_degree_core.py
@@ -13,7 +13,7 @@ from typing import Any
 import numpy as np
 import numpy.typing as npt
 
-from ._numba_compat import nb_jit
+from .._numba_compat import nb_jit
 
 
 @nb_jit(nopython=True, cache=True)
