+++ src/pantr/bspline/_bspline_space_factory.py
