+++ tests/test_bspline.py
@@ -4,8 +4,8 @@ import numpy as np
 import numpy.typing as npt
 import pytest
 
-from pantr._bspline_basis_core import _compute_basis_nurbs_book_impl
 from pantr.bspline import Bspline, BsplineSpace, BsplineSpace1D, create_uniform_periodic
+from pantr.bspline._bspline_basis_core import _compute_basis_nurbs_book_impl
 
 
 class TestBsplineInit:
