+++ tests/test_bspline_basis_derivatives_1D.py
@@ -10,13 +10,13 @@ from pantr._basis_core import (
     _tabulate_Bernstein_basis_1D_core,
     _tabulate_Bernstein_basis_deriv_1D_core,
 )
-from pantr._bspline_basis_core import (
+from pantr.bspline import BsplineSpace1D
+from pantr.bspline._bspline_basis_core import (
     _compute_basis_deriv_nurbs_book_impl,
     _compute_basis_nurbs_book_impl,
     _tabulate_Bspline_basis_Bernstein_like_deriv_1D,
     _tabulate_Bspline_basis_deriv_1D_impl,
 )
-from pantr.bspline import BsplineSpace1D
 
 # ---------------------------------------------------------------------------
 # Shared fixtures
