+++ tests/test_bspline_product.py
@@ -6,9 +6,9 @@ import numpy as np
 import numpy.typing as npt
 import pytest
 
-from pantr._bspline_basis_core import _compute_basis_nurbs_book_impl
-from pantr._bspline_knots import _get_unique_knots_and_multiplicity_impl
 from pantr.bspline import Bspline, BsplineSpace, BsplineSpace1D, create_uniform_periodic
+from pantr.bspline._bspline_basis_core import _compute_basis_nurbs_book_impl
+from pantr.bspline._bspline_knots import _get_unique_knots_and_multiplicity_impl
 
 # ---------------------------------------------------------------------------
 # Helpers
