+++ tests/test_bspline_product_nd.py
@@ -10,8 +10,8 @@ from pantr._bezier_product import (
     _bernstein_product_coefficients,
     _bernstein_product_coefficients_nd,
 )
-from pantr._bspline_knots import _get_unique_knots_and_multiplicity_impl
 from pantr.bspline import Bspline, BsplineSpace, BsplineSpace1D, create_uniform_periodic
+from pantr.bspline._bspline_knots import _get_unique_knots_and_multiplicity_impl
 from pantr.quad import PointsLattice
 
 # ---------------------------------------------------------------------------
